Move some stuff to common namespaces(targets). Start combining consumer with compNode service.
