--- computationalLib/lib/computationalLib/interfaces/DataAcessInteface.cpp
@@ -1,2 +1,0 @@
-
-#include "computationalLib/interfaces/DataAcessInteface.h"
