+++ computationalLib/lib/computationalLib/network_shared/restUtils.cpp
@@ -1,4 +1,4 @@
-#include "restUtils.h"
+#include "network_shared/restUtils.h"
 
 size_t WriteCallback(void *contents,
                      size_t size,
