+++ computationalLib/test/math_core/smokeTests.cpp
@@ -2,7 +2,7 @@
 #include <chrono>
 
 #include "common/sharedDeclarations.h"
-#include "computationalLib/math_core/Dipoles.h"
+#include "math_core/Dipoles.h"
 #include "common/commonTypes.h"
 #include "common/Parsers.h"
 #include "common/Printers.h"
