+++ computationalNode/CMakeLists.txt
@@ -18,7 +18,7 @@ file(GLOB_RECURSE httpSources CONFIGURE_DEPENDS
 
 
 add_library(servLib ${httpSources})
-target_link_libraries(servLib PUBLIC math_core_lib Drogon::Drogon)
+target_link_libraries(servLib PUBLIC math_core_lib Drogon::Drogon network_shared_lib)
 target_include_directories(servLib PUBLIC include)
 enable_coverage_for_target(servLib)
