+++ computationalNode/src/service/CompNodeService.cpp
@@ -1,6 +1,3 @@
 #include "service/CompNodeService.h"
 
-void CompNodeService::asyncHandleHttpRequest(const HttpRequestPtr &req,
-                                             std::function<void(const HttpResponsePtr &)> &&callback) {
-  // write your application logic here
-}
+
