+++ mainNode/CMakeLists.txt
@@ -33,15 +33,15 @@ target_link_libraries(mainNode -lssl -lcrypto)
 target_link_libraries(mainNode servLib2 allUtils)
 enable_coverage_for_target(mainNode)
 
-add_library(amqpUtils src/amqpCommon.cpp src/amqpRestService.cpp src/restUtils.cpp)
-target_link_libraries(amqpUtils PUBLIC fmt::fmt amqpcpp jsoncpp_lib CURL::libcurl)
+#[[add_library(amqpUtils ../computationalLib/lib/computationalLib/network_shared/amqpCommon.cpp ../computationalLib/lib/computationalLib/network_shared/amqpRestService.cpp ../computationalLib/lib/computationalLib/network_shared/restUtils.cpp)
+target_link_libraries(amqpUtils PUBLIC  network_shared_lib)
 target_include_directories(amqpUtils PUBLIC ../utilLib/include)
 enable_coverage_for_target(amqpUtils)
-target_include_directories(amqpUtils PUBLIC include)
+target_include_directories(amqpUtils PUBLIC include)]]
 
 add_executable(consumer aqmpCons.cpp)
-target_link_libraries(consumer PUBLIC amqpUtils)
+target_link_libraries(consumer PUBLIC network_shared_lib)
 
 
 add_executable(producer aqmpProd.cpp)
-target_link_libraries(producer PUBLIC amqpUtils)
+target_link_libraries(producer PUBLIC network_shared_lib)
